basic implementation of page
