+++ react_ui/src/components/AnimatedRoutes/AnimatedRoutes.tsx
@@ -4,7 +4,7 @@ import { AnimatePresence } from 'framer-motion';
 import ResumePage from '../../scenes/pages/ResumePage';
 import NotesPage from '../../scenes/pages/Notes/NotesPage';
 import ContactPage from '../../scenes/pages/ContactPage';
-import HomePage from '../../scenes/pages/HomePage';
+import HomePage from '../../scenes/pages/Home/HomePage';
 import AboutPage from '../../scenes/pages/AboutPage';
 
 const AnimatedRoutes = () => {
