+++ react_ui/src/index.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import './scenes/pages/Notes/NotesPage.css';
+import './scenes/pages/Home/HomePage.css';
 import App from './App';
 
 const root = ReactDOM.createRoot(
