+++ react_ui/src/scenes/pages/AboutPage.tsx
@@ -39,6 +39,21 @@ const AboutPage = () => {
       </div>
       <div className='flex justify-content-center align-items-center'>
         <p>{data.bio}</p>
+        <p>
+          Hey there, I'm Jacob, a Senior Software Engineer with a passion for creating customer-oriented products that meet user needs. I have expertise in developing full-stack web applications, architecting database systems, and implementing security measures to keep data safe.
+        </p>
+        <p>
+          Throughout my career, I have worked with a variety of programming languages such as Ruby, Kotlin, C#, and Java, and have experience working with popular frameworks such as Ruby on Rails, React, and .NET Core. I am also knowledgeable in different architectures such as MVC, microservices, and service-oriented architecture.
+        </p>
+        <p>
+          What drives me is a passion for creating products that meet the needs of both customers and businesses. I believe that well-designed and well-implemented software can revolutionize the way people work and live. I strive to create software that is not only functional and efficient but also user-friendly and accessible to everyone.
+        </p>
+        <p>
+          I have experience in leading teams of developers, mentoring junior developers, and working closely with product managers to ensure the development process aligns with business goals. I have also implemented security measures in applications, using automated security suites to monitor and address application security threats.
+        </p>
+        <p>
+          If you're looking for a senior software engineer who can hit the ground running and work collaboratively to achieve your goals, let's chat.
+        </p>
       </div>
       <div className='flex justify-content-center align-items-center'>
         <div className='flex flex-column'>
