+++ react_ui/src/scenes/pages/Home/HomePage.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { Box } from '@mui/material';
+
+const HomePage = () => {
+  return (
+    <Box className='home__container'>
+      <Box className='home__name'>Jacob Penner</Box>
+      <Box className='home__title'>Full Stack Software Developer</Box>
+    </Box>
+  )
+};
+
+export default HomePage;
